Minor Changes to WEBSERVICE
The documentation on how to configure an Http client was out of date. There were some minor flaws in the implementation, including the fact that you either had to configure a client or face an exception if you were evaluating cells which used the function (e.g. when saving a spreadsheet that someone else may have provided). A non-default setting is provided to treat the result as `#VALUE!` if this situation arises.